refactor: move validateGroups allAgesCovered functions out
